% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse.R
\name{hsl_to_rgb}
\alias{hsl_to_rgb}
\title{HSL to sRGB}
\usage{
hsl_to_rgb(H, S, L)
}
\arguments{
\item{H}{Hue in 0-360 degrees.}

\item{S, L}{Saturation and lightness, each 0-1.}
}
\value{
A named list of three matrices `R`, `G`, `B` in **0-1**.
}
\description{
HSL to sRGB
}
\details{
`H` is not wrapped; see [hsv_to_rgb()].
}
\examples{
hsl <- rgb_to_hsl(matrix(200, 1, 1), matrix(100, 1, 1), matrix(50, 1, 1))
round(hsl_to_rgb(hsl$H, hsl$S, hsl$L)$R * 255, 2)
}
