% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_rgb}
\alias{read_rgb}
\title{Read the RGB bands of a raster}
\usage{
read_rgb(x)
}
\arguments{
\item{x}{A `SpatRaster` or a path to a raster file. Bands 1-3 are taken
as R, G and B; further bands are ignored.}
}
\value{
A list with matrices `R`, `G`, `B` and the `template`
  `SpatRaster` the geometry came from.
}
\description{
Read the RGB bands of a raster
}
\details{
Requires \pkg{terra}, which is only a Suggests: every conversion in this
package works on plain matrices and has no spatial dependency at all.
}
\examples{
\donttest{
if (requireNamespace("terra", quietly = TRUE)) {
  r <- terra::rast(nrows = 4, ncols = 4, nlyrs = 3, vals = runif(48, 0, 255))
  bands <- read_rgb(r)
  dim(bands$R)
}
}
}
